requirements.txt
openai
speechrecognition
pyttsx3
sounddevice
numpy
python-dotenv
